refactor: move vfn factory into module
